TextMonitoring + Tests

+++ src/main/java/com/rakovets/course/javabasics/example/collections/map/TextMonitoring.java
@@ -7,6 +7,7 @@ import java.util.Map;
 public final class TextMonitoring {
 
     public static HashMap<String, Integer> researchText(String text) {
+
         String[] textArray = text.split("[^a-zA-Z]+");
         HashMap<String, Integer> textMap = new HashMap<>();
         for (String word : textArray) {
@@ -22,16 +23,15 @@ public final class TextMonitoring {
     }
 
     public static int getCountUniqueWords(String text) {
+
         HashMap<String, Integer> newmap = researchText(text);
         int count = 0;
         for (Map.Entry<String, Integer> item : newmap.entrySet()) {
             if (item.getValue() == 1) {
                 count++;
             }
-
         }
         return count;
-
     }
 
     public static LinkedList<String> getUniqueWords(String text) {
@@ -44,16 +44,15 @@ public final class TextMonitoring {
             }
         }
         return result;
-
     }
 
     public static int getWork(String text, String checkingWord) {
-   int result = 0;
+
+        int result = 0;
         for (Map.Entry<String, Integer> item : researchText(text).entrySet()) {
             if (item.getKey().equals(checkingWord)) {
-                 result = (int)item.getValue();
+                result = item.getValue();
             }
-
         }
         return result;
     }
